chore: merge models
