+++ src/renderer/features/multisig-operations/components/modals/AccountNotFoundModal.tsx
@@ -4,13 +4,13 @@ import { useTranslation } from 'react-i18next';
 import { Animation } from '@/shared/ui/Animation/Animation';
 import { Button } from '@/shared/ui/Buttons/Button/Button';
 import { StatusModal } from '@/shared/ui/Modals/StatusModal/StatusModal';
-import { accountNotFoundModal } from '../../deep-link/multisig-operation-deep-link';
+import { deepLinkModel } from '../../model/deep-link';
 
 export const AccountNotFoundModal = () => {
   const { t } = useTranslation();
 
-  const isOpen = useUnit(accountNotFoundModal.$isOpen);
-  const handleClose = () => accountNotFoundModal.close();
+  const isOpen = useUnit(deepLinkModel.$isAccountNotFoundModalOpen);
+  const handleClose = () => deepLinkModel.closeNotFoundModal();
 
   return (
     <StatusModal
